+++ gen/proto/go/parca/metastore/v1alpha1/metastore.pb.gw.go
@@ -633,7 +633,7 @@ func RegisterMetastoreServiceHandlerServer(ctx context.Context, mux *runtime.Ser
 // RegisterMetastoreServiceHandlerFromEndpoint is same as RegisterMetastoreServiceHandler but
 // automatically dials to "endpoint" and closes the connection when "ctx" gets done.
 func RegisterMetastoreServiceHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
-	conn, err := grpc.Dial(endpoint, opts...)
+	conn, err := grpc.DialContext(ctx, endpoint, opts...)
 	if err != nil {
 		return err
 	}