+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.115 (2023-03-20)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.114](https://github.com/parca-dev/parca/compare/@parca/components@0.16.113...@parca/components@0.16.114) (2023-03-16)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.113 (2023-03-16)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.112 (2023-03-15)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.111](https://github.com/parca-dev/parca/compare/@parca/components@0.16.110...@parca/components@0.16.111) (2023-03-13)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.110](https://github.com/parca-dev/parca/compare/@parca/components@0.16.109...@parca/components@0.16.110) (2023-03-07)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.109](https://github.com/parca-dev/parca/compare/@parca/components@0.16.108...@parca/components@0.16.109) (2023-03-07)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.108](https://github.com/parca-dev/parca/compare/@parca/components@0.16.107...@parca/components@0.16.108) (2023-03-07)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.107 (2023-03-07)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.106 (2023-03-02)
+
+# 0.16.0 (2023-03-01)
+
+**Note:** Version bump only for package @parca/components
+
+## [0.16.105](https://github.com/parca-dev/parca/compare/@parca/components@0.16.104...@parca/components@0.16.105) (2023-02-27)
+
+**Note:** Version bump only for package @parca/components
+
 ## 0.16.104 (2023-02-23)
 
 **Note:** Version bump only for package @parca/components